put seed_everything into the utils file. also rename the utils file
