+++ geometric_utils.py
@@ -1,4 +1,5 @@
 # for a single set of poin clouds (one training example), convert it into a graph
+import math
 import numpy as np
 
 from irrep import Irreps
@@ -55,4 +56,5 @@ def avg_irreps_with_same_id(irreps_list: list[Irreps]) -> Irreps:
 
     for i in range(len(summed_data)):
         summed_data[i] /= len(irreps_list)
+        # summed_data[i] *= math.sqrt(1 / len(irreps_list))
     return Irreps.from_id(irreps_list[0].id(), summed_data)
